Update onetime from nitime.
